Added a display warning messagebox when the user tries to delte an product that has associated parts with that product. User must modify the productt and remove those associated parts before having access grant for the deletetion. Modified the [ModifyProductForm] save button event to fit this scenerio. It presviously forced the user to have at least one associated part before saving. That functionality remains in the [AddProductForm].

+++ Inventory Management System (WinForm)/View/ModifcationControlsUC.cs
@@ -131,8 +131,21 @@ namespace Inventory_Managment_System.View
 
                     if (confirmedDeletion)
                     {
-                        Controller.Controller.deleteProductFromInventory(selectedProductsRowIndex);
-                        UIDataGridViewValidator.recreateTableData(productsDataGridView, Inventory.Products);
+                        var productAssociatedParts_Count = Controller.Controller.lookUpProductFromTheInventory(selectedProductsRowIndex).AssociatedParts.Count;
+                        
+                        try
+                        {
+                            if (productAssociatedParts_Count == 0)
+                            {
+                                Controller.Controller.deleteProductFromInventory(selectedProductsRowIndex);
+                                UIDataGridViewValidator.recreateTableData(productsDataGridView, Inventory.Products);
+                            } else
+                            {
+                                throw new ConstraintException();
+                            }
+                        } catch(ConstraintException) {
+                            UIMsgBox.displayDeletingProductWithAssociatedParts(productAssociatedParts_Count);
+                        }
                     }
                 } else
                 {
